requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings== 2.7.1
aiofiles==23.2.1
langchain_community ==0.3.17
pymupdf == 1.25.3



